SpeechToSign/requirements.txt
speechrecognition>=3.10.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pyaudio>=0.2.11
